+++ AgriMarket.Web/Areas/Client/ViewModels/Bookings/BookingIndexItemViewModel.cs
@@ -8,6 +8,6 @@ public class BookingIndexItemViewModel
     public string ListingTitle { get; set; } = default!;
     public BookingStatus Status { get; set; }
     public decimal TotalPrice { get; set; }
-    public double AreaInHectares { get; set; }
+    public decimal AreaInHectares { get; set; }
     public DateTime CreatedAt { get; set; }
 }
